+++ shared/src/commonMain/kotlin/ui/composables/credentials/CredentialCardHeader.kt
@@ -16,8 +16,8 @@ import ui.composables.PersonAttributeDetailCardHeadingIcon
 @Composable
 fun ColumnScope.CredentialCardHeader(
     credential: SubjectCredentialStore.StoreEntry,
+    showLoadingSpinner: Boolean,
     onDelete: () -> Unit,
-    modifier: Modifier = Modifier,
 ) {
     PersonAttributeDetailCardHeading(
         icon = {
@@ -31,6 +31,7 @@ fun ColumnScope.CredentialCardHeader(
         },
     ) {
         CredentialCardActionMenu(
+            showLoadingSpinner = showLoadingSpinner,
             onDelete = onDelete
         )
     }