+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Release 5.5.3
- - tbd
+ * Display credential validity for invalid credentials
 
 # Release 5.5.2
  * Fix presentation of ISO credentials